Fix ReCaptch insjection in all drivers
