+++ src/OrchardCore.Modules/OrchardCore.ReCaptcha/Services/ReCaptchaUserRegistrationShapeTableProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace OrchardCore.ReCaptcha.Services;
+
+internal sealed class ReCaptchaUserRegistrationShapeTableProvider : ReCaptchaShapeTableProvider
+{
+    public ReCaptchaUserRegistrationShapeTableProvider()
+        : base("RegisterUserForm_Edit")
+    {
+    }
+}
