+++ src/OrchardCore.Modules/OrchardCore.ReCaptcha/Services/ReCaptchaUserResetPasswordShapeTableProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace OrchardCore.ReCaptcha.Services;
+
+internal sealed class ReCaptchaUserResetPasswordShapeTableProvider : ReCaptchaShapeTableProvider
+{
+    public ReCaptchaUserResetPasswordShapeTableProvider()
+        : base("ResetPasswordForm_Edit", "ForgotPasswordForm_Edit")
+    {
+    }
+}
