+++ src/OrchardCore.Modules/OrchardCore.ReCaptcha/Services/ReCaptchaUserShapeTableProvider.cs
@@ -1,0 +1,9 @@
+namespace OrchardCore.ReCaptcha.Services;
+
+internal sealed class ReCaptchaUserShapeTableProvider : ReCaptchaShapeTableProvider
+{
+    public ReCaptchaUserShapeTableProvider()
+        : base("LoginForm_Edit")
+    {
+    }
+}
