+++ src/OrchardCore.Modules/OrchardCore.ReCaptcha/Startup.cs
@@ -12,7 +12,6 @@ using OrchardCore.Security.Permissions;
 using OrchardCore.Settings.Deployment;
 using OrchardCore.Users;
 using OrchardCore.Users.Events;
-using OrchardCore.Users.Models;
 
 namespace OrchardCore.ReCaptcha;
 
@@ -46,7 +45,7 @@ public sealed class UsersStartup : StartupBase
         services.AddScoped<IRegistrationFormEvents, RegistrationFormEventHandler>();
         services.AddScoped<ILoginFormEvent, LoginFormEventEventHandler>();
         services.AddScoped<IPasswordRecoveryFormEvents, PasswordRecoveryFormEventEventHandler>();
-        services.AddShapeTableProvider<ReCaptchaShapeTableProvider>();
+        services.AddShapeTableProvider<ReCaptchaUserShapeTableProvider>();
     }
 }
 
@@ -56,8 +55,7 @@ public sealed class UsersResetPasswordStartup : StartupBase
 {
     public override void ConfigureServices(IServiceCollection services)
     {
-        services.AddDisplayDriver<ForgotPasswordForm, ReCaptchaForgotPasswordFormDisplayDriver>();
-        services.AddDisplayDriver<ResetPasswordForm, ReCaptchaResetPasswordFormDisplayDriver>();
+        services.AddShapeTableProvider<ReCaptchaUserResetPasswordShapeTableProvider>();
     }
 }
 
@@ -67,6 +65,6 @@ public sealed class UsersRegistrationStartup : StartupBase
 {
     public override void ConfigureServices(IServiceCollection services)
     {
-        services.AddDisplayDriver<RegisterUserForm, RegisterUserFormDisplayDriver>();
+        services.AddShapeTableProvider<ReCaptchaUserRegistrationShapeTableProvider>();
     }
 }
