+++ src/OrchardCore.Themes/TheAdmin/Assets/scss/components/_recaptcha.scss
@@ -1,0 +1,3 @@
+.g-recaptcha {
+    margin-bottom: 1rem;
+}
