+++ src/OrchardCore.Themes/TheAdmin/Assets/scss/index.scss
@@ -20,6 +20,7 @@
 @import 'components/navbar';
 @import 'components/nouislider';
 @import 'components/pager';
+@import 'components/recaptcha';
 @import 'components/sortable';
 @import 'components/tabs';
 @import 'components/trumbowyg';
