+++ src/OrchardCore.Themes/TheAdmin/Assets/scss/login.scss
@@ -8,4 +8,5 @@ body, html {
     margin: auto;
 }
 
+@import 'components/recaptcha';
 @import 'components/validations';
