+++ src/OrchardCore.Themes/TheAdmin/wwwroot/css/TheAdmin.css
@@ -8461,6 +8461,10 @@ ul.pager {
   justify-content: center;
 }
 
+.g-recaptcha {
+  margin-bottom: 1rem;
+}
+
 /* 
     IMPORTANT: Never import Bootstrap directly into the theme. 
     TheAdmin.css will depend on Bootstrap, but we never want to compile it.
