+++ src/OrchardCore.Themes/TheAgencyTheme/Assets/src/scss/bootstrap-oc.scss
@@ -11,6 +11,7 @@ By Orchard Core Team
 @import "modules/_flow.scss";
 @import "modules/_messages.scss";
 @import "modules/_widgets.scss";
+@import "modules/_recaptcha.scss";
 @import "modules/_pager.scss";
 @import "modules/_taxonomy.scss";
