+++ src/OrchardCore.Themes/TheBlogTheme/Assets/src/scss/bootstrap-oc.scss
@@ -12,6 +12,7 @@ By Orchard Core Team
 @import "modules/_messages.scss";
 @import "modules/_widgets.scss";
 @import "modules/_pager.scss";
+@import "modules/_recaptcha.scss";
 @import "modules/_taxonomy.scss";
 
 #togglePassword {
