+++ src/OrchardCore.Themes/TheBlogTheme/wwwroot/css/bootstrap-oc.css
@@ -253,6 +253,10 @@ ul.pager {
   margin-top: 1rem;
 }
 
+.g-recaptcha {
+  margin-bottom: 1rem;
+}
+
 .term-badge > a {
   text-decoration: none;
 }
