+++ src/OrchardCore.Themes/TheTheme/Assets/scss/modules/_recaptcha.scss
@@ -1,0 +1,3 @@
+.g-recaptcha {
+    margin-bottom: 1rem;
+}
