+++ src/OrchardCore.Themes/TheTheme/Assets/scss/theme.scss
@@ -3,6 +3,7 @@
 
 /* Include modules */
 @import "modules/pager";
+@import "modules/recaptcha";
 @import "modules/taxonomy";
 @import "modules/widgets";
