+++ src/docs/releases/3.0.0.md
@@ -102,19 +102,49 @@ will return `false` for administrators, even though they still have full access.
 {% assign isAuthorized = User | has_permission: "AccessAdminPanel" %}
 ```
 
-### Sealing Types
+### ReCaptcha
 
-Many type commonly used by classes can be `sealed`, which improves runtime performance. While it's not mandatory, we recommend that you consider applying this improvement to your own extensions as well. We've implemented this enhancement in pull request [#16897](https://github.com/OrchardCMS/OrchardCore/pull/16897).
+Previously, the ReCaptcha module included a shape type called "FormReCaptcha," which allowed you to use the driver to display the ReCaptcha challenge on any form. However, due to the dynamic nature of ReCaptcha and the threshold settings that control when the challenge is displayed, it is crucial to inject ReCaptcha using the `ShapeTableProvider`.
+
+If you're currently using `FormReCaptcha` in your project, you should transition to injecting it through an implementation of `IShapeTableProvider`. To simplify this process, we’ve introduced a new base class, `ReCaptchaShapeTableProvider`, which automatically injects the ReCaptcha challenge for specific shape types. For example, we inject it on the login page in Orchard Core as follows:
+
+```csharp
+services.AddShapeTableProvider<ReCaptchaUserShapeTableProvider>();
+```
+
+The `ReCaptchaUserShapeTableProvider` class inherits from the new `ReCaptchaShapeTableProvider`, which you can also use directly:
+
+```csharp
+internal sealed class ReCaptchaUserShapeTableProvider : ReCaptchaShapeTableProvider
+{
+    public ReCaptchaUserShapeTableProvider()
+        : base("LoginForm_Edit")
+    {
+    }
+}
+```
+
+Alternatively, if you must continue rendering ReCaptcha using the deprecated `FormReCaptcha`, you can manually add the `FormReCaptcha.cshtml` view to your project with the following code:
+
+```html
+<div class="mb-3">
+    <captcha mode="AlwaysShow" language="@Orchard.CultureName()" />
+</div>
+```
 
 ## Change Log
 
-### reCaptcha
+### ReCaptcha
 
 ### New ReCaptcha Shape
 
 A new `ReCaptcha` shape has been introduced, enabling you to render the ReCaptcha challenge using a customizable shape. For more details, please refer to the [documentation](../reference/modules/ReCaptcha/README.md).
 
 ## Miscellaneous
+
+### Sealing Types
+
+Many type commonly used by classes can be `sealed`, which improves runtime performance. While it's not mandatory, we recommend that you consider applying this improvement to your own extensions as well. We've implemented this enhancement in pull request [#16897](https://github.com/OrchardCMS/OrchardCore/pull/16897).
 
 ### New Extension Method for Adding `IConfigureOptions<ResourceManagementOptions>` Implementations
